Gauge conf parallelized
