+++ include/config.h
@@ -1,5 +1,5 @@
 #ifndef CONFIG_H
 #define CONFIG_H
-#define NS 64
-#define NT 64
+#define NS 16
+#define NT 16
 #endif
