+++ src/main.cpp
@@ -2,107 +2,84 @@
 #include <ctime>
 #include <fstream>
 #include <string>
-#include "hmc.h"
+#include "conjugate_gradient.h"
 #include <format>
 
+//Formats decimal numbers
+//For opening file with confs 
+static std::string format(const double& number) {
+    std::ostringstream oss;
+    oss << std::fixed << std::setprecision(4) << number;
+    std::string str = oss.str();
+    str.erase(str.find('.'), 1); //Removes decimal dot 
+    return str;
+}
 
-int main() {
+int main(int argc, char **argv) {
+    MPI_Init(&argc, &argv);
+    MPI_Comm_size(MPI_COMM_WORLD, &mpi::size);
+    MPI_Comm_rank(MPI_COMM_WORLD, &mpi::rank);
+    mpi::maxSize = (mpi::rank != mpi::size-1) ? LV::Ntot/mpi::size :  LV::Ntot/mpi::size + LV::Ntot%mpi::size;
+
+
+
     allocate_lattice_arrays();
-    srand(time(0));
+
     
-    int Ntherm, Nmeas, Nsteps, Nm0; //Simulation parameters
-    double beta; //Beta range
-    double trajectory_length; //HMC parameters
-    int MD_steps;
-    double m0_min, m0_max; //bare mass
-	int saveconf = 0; //Save configurations
+
+    srand(mpi::rank*19);
+
+    double beta = 2; 
+    double m0 = -0.18840579710144945;
+    int nconf=0;
+    if (LV::Nx == 64)
+        nconf = 0;
+    else if (LV::Nx == 128)
+        nconf = 3;
+    else if (LV::Nx == 256)
+        nconf = 20;  
 
     CG::max_iter = 10000; //Maximum number of iterations for the conjugate gradient method
     CG::tol = 1e-10; //Tolerance for convergence
-    //---Input data---//
-    std::cout << "  -----------------------------" << std::endl;
-    std::cout << "|  Two-flavor Schwinger model   |" << std::endl;
-    std::cout << "| Hybrid Monte Carlo simulation |" << std::endl;
-    std::cout << "  -----------------------------" << std::endl;
-    std::cout << "Nx " << LV::Nx << " Nt " << LV::Nt << std::endl;
-    std::cout << "m0 min: ";
-    std::cin >> m0_min;
-    std::cout << "m0 max: ";
-    std::cin >> m0_max;
-    std::cout << "Number of masses in [m0_min, m0_max]: ";
-    std::cin >> Nm0;
-    std::cout << "Molecular dynamics steps: ";
-    std::cin >> MD_steps;
-    std::cout << "Trajectory length: ";
-    std::cin >> trajectory_length; 
-    std::cout << "beta: ";
-    std::cin >> beta;
-    std::cout << "Thermalization: ";
-    std::cin >> Ntherm;
-    std::cout << "Measurements: ";
-    std::cin >> Nmeas;
-    std::cout << "Step (sweeps between measurements): ";
-    std::cin >> Nsteps;
-    std::cout << "Save configurations yes/no (1 or 0): ";
-    std::cin >> saveconf;
-    std::cout << " " << std::endl;
     
-    std::vector<double> Masses(Nm0);
-
     periodic_boundary(); //Compute right and left periodic boundary
     
-	GaugeConf GConf = GaugeConf(LV::Nx, LV::Nt);  //Gauge configuration
- 
-    if (Nm0 == 1) {
-        Masses = { m0_min };
-    }
-    else {
-        Masses = linspace(m0_min, m0_max, Nm0);
-    }
+	GaugeConf GConf = GaugeConf();  //Gauge configuration
+    GConf.initialization(); //Random initialization of the gauge configuration 
 
     std::ostringstream NameData;
-    NameData << "2D_U1_Ns" << LV::Nx << "_Nt" << LV::Nt << "_Meas" << Nmeas << ".txt";
-    std::ofstream Datfile;
-    Datfile.open(NameData.str());
-    Datfile << std::format("{:<30.17g}{:<30d}{:<30d}{:<30d}\n", beta, Ntherm, Nmeas, Nsteps);
-    Datfile << std::format("{:<30.17g}{:<30d}\n", trajectory_length, MD_steps);
-    for (double m0 : Masses) {
-        std::cout << "**********************************************************************" << std::endl;
-        std::cout << "*                              PARAMETERS" << std::endl;
-        std::cout << "* Nx = " << LV::Nx << ", Nt = " << LV::Nt << std::endl;
-        std::cout << "* m0 = " << m0 << ", kappa = " << 1/(2*(m0+2)) << std::endl;
-        std::cout << "* beta = " << beta << std::endl;
-        std::cout << "* Thermalization confs = " << Ntherm << std::endl;
-        std::cout << "* Measurement confs = " << Nmeas << std::endl;
-        std::cout << "* Decorrelation steps (confs dropped between measurements) = " << Nsteps << std::endl;
-        std::cout << "* Trajectory length = " << trajectory_length << ", Leapfrog steps = " << MD_steps << 
-        ", Integration step = " << trajectory_length/MD_steps << std::endl;
-        std::cout << "* CG max iterations = " << CG::max_iter << ", CG tolerance = " << CG::tol << std::endl;
-        std::cout << "**********************************************************************" << std::endl;
-
-
-        HMC hmc = HMC(GConf,MD_steps, trajectory_length, Ntherm, Nmeas, Nsteps, beta, LV::Nx, LV::Nt, m0,saveconf);   
-        double begin = omp_get_wtime();
-        hmc.HMC_algorithm();
-        double end = omp_get_wtime();
-        std::cout << "Average plaquette value / volume: Ep = " << hmc.getEp() << " dEp = " << hmc.getdEp() << std::endl;
-        std::cout << "Average gauge action / volume: gS = " << hmc.getgS() << " dgS = " << hmc.getdgS() << std::endl;
-        std::cout << "Acceptance rate: " << hmc.getacceptance_rate() << std::endl;
-        double elapsed_secs = end - begin;
-
-        Datfile << std::format("{:<30.17g}\n", m0);
-        Datfile << std::format("{:<30.17g}{:<30.17g}\n", hmc.getEp(), hmc.getdEp());
-        Datfile << std::format("{:<30.17g}{:<30.17g}\n", hmc.getgS(), hmc.getdgS());
-        Datfile << std::format("{:<30.17g}\n", hmc.getacceptance_rate());
-        Datfile << std::format("{:<30.17g}", elapsed_secs);
-        
+    NameData << "../confs/b" << beta << "_" << LV::Nx << "x" << LV::Nt << "/m-018/2D_U1_Ns" << LV::Nx << "_Nt" << LV::Nt << "_b" << 
+    format(beta).c_str() << "_m" << format(m0).c_str() << "_" << nconf << ".ctxt";
+    GConf.read_conf(NameData.str());
  
-        std::cout << "Time = " << elapsed_secs << " s" << std::endl;
-        std::cout << "-------------------------------" << std::endl;
+    spinor sol(mpi::maxSize), rhs(mpi::maxSize);
 
+    for(int n = 0; n <mpi::maxSize; n++) {
+        rhs.mu0[n] = 1;//RandomU1(); //spin up
+        rhs.mu1[n] = 1;//RandomU1(); //spin down
     }
-    Datfile.close();
+  
+    D_phi(GConf.Conf, rhs, sol, m0); //Applies Dirac operator to rhs and stores the result in sol
+
+    //print sol for each rank
     
+    for(int i = 0; i < mpi::size; i++) {
+        MPI_Barrier(MPI_COMM_WORLD);
+        if (i == mpi::rank) {
+            printf("Rank %d\n", mpi::rank);
+            for(int n = 0; n < mpi::maxSize; n++) {
+                std::cout << "Rank " << mpi::rank << " sol.mu0[" << n << "] = " << sol.mu0[n] << std::endl;
+                std::cout << "Rank " << mpi::rank << " sol.mu1[" << n << "] = " << sol.mu1[n] << std::endl;
+            }
+        }
+    }
+    
+
+    MPI_Finalize();
+
+
+    
+
     //Free coordinate arrays
     free_lattice_arrays();
